Rename component to event publisher proxy  (#9475)
