--- components/cloud-event-gateway-proxy/config/100-secret.yaml
@@ -1,11 +1,0 @@
-apiVersion: v1
-data:
-  client-id: fake-id
-  client-secret: fake-secret
-  token-endpoint: fake-token-ep
-  ems-publish-url: fake-ems-publish-url
-kind: Secret
-metadata:
-  labels:
-    app: cloud-event-gateway-proxy
-  name: cloud-event-gateway-proxy
