--- components/cloud-event-gateway-proxy/config/200-service.yaml
@@ -1,12 +1,0 @@
-apiVersion: v1
-kind: Service
-metadata:
-  name: cloud-event-gateway-proxy
-spec:
-  type: ClusterIP
-  selector:
-    app: cloud-event-gateway-proxy
-  ports:
-    - protocol: TCP
-      port: 80
-      targetPort: 8080
