--- components/cloud-event-gateway-proxy/config/300-deployment.yaml
@@ -1,63 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  labels:
-    app: cloud-event-gateway-proxy
-  name: cloud-event-gateway-proxy
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      app: cloud-event-gateway-proxy
-  strategy:
-    type: RollingUpdate
-  template:
-    metadata:
-      labels:
-        app: cloud-event-gateway-proxy
-    spec:
-      containers:
-        - env:
-            - name: CLIENT_ID
-              valueFrom:
-                secretKeyRef:
-                  name: cloud-event-gateway-proxy
-                  key: client-id
-            - name: CLIENT_SECRET
-              valueFrom:
-                secretKeyRef:
-                  name: cloud-event-gateway-proxy
-                  key: client-secret
-            - name: TOKEN_ENDPOINT
-              valueFrom:
-                secretKeyRef:
-                  name: cloud-event-gateway-proxy
-                  key: token-endpoint
-            - name: EMS_PUBLISH_URL
-              valueFrom:
-                secretKeyRef:
-                  name: cloud-event-gateway-proxy
-                  key: ems-publish-url
-          image: github.com/kyma-project/kyma/components/cloud-event-gateway-proxy/cmd/cloud-event-gateway-proxy
-          imagePullPolicy: IfNotPresent
-          name: cloud-event-gateway-proxy
-          ports:
-            - containerPort: 8080
-              name: http
-              protocol: TCP
-          livenessProbe:
-            failureThreshold: 3
-            httpGet:
-              path: /healthz
-              port: 8080
-              scheme: HTTP
-            initialDelaySeconds: 5
-            periodSeconds: 2
-            successThreshold: 1
-            timeoutSeconds: 1
-          readinessProbe:
-            failureThreshold: 3
-            httpGet:
-              path: /readyz
-              port: 8080
-              scheme: HTTP
