--- components/cloud-event-gateway-proxy/go.mod
@@ -1,12 +1,0 @@
-module github.com/kyma-project/kyma/components/cloud-event-gateway-proxy
-
-go 1.15
-
-require (
-	github.com/cloudevents/sdk-go/v2 v2.2.0
-	github.com/kelseyhightower/envconfig v1.4.0
-	github.com/pkg/errors v0.9.1
-	github.com/sirupsen/logrus v1.6.0
-	go.opencensus.io v0.22.4
-	golang.org/x/oauth2 v0.0.0-20200107190931-bf48bf16ab8d
-)
