+++ components/event-publisher-proxy/config/100-secret.yaml
@@ -1,0 +1,11 @@
+apiVersion: v1
+data:
+  client-id: fake-id
+  client-secret: fake-secret
+  token-endpoint: fake-token-ep
+  ems-publish-url: fake-ems-publish-url
+kind: Secret
+metadata:
+  labels:
+    app: event-publisher-proxy
+  name: event-publisher-proxy
