+++ components/event-publisher-proxy/config/200-service.yaml
@@ -1,0 +1,12 @@
+apiVersion: v1
+kind: Service
+metadata:
+  name: event-publisher-proxy
+spec:
+  type: ClusterIP
+  selector:
+    app: event-publisher-proxy
+  ports:
+    - protocol: TCP
+      port: 80
+      targetPort: 8080
