+++ components/event-publisher-proxy/testing/fixtures.go
